+++ packages/browser/src/sdk.ts
@@ -2,6 +2,7 @@ import { getCurrentHub, initAndBind, Integrations as CoreIntegrations } from '@s
 
 import { BrowserOptions } from './backend';
 import { BrowserClient, ReportDialogOptions } from './client';
+import { wrap as internalWrap } from './helpers';
 import { Breadcrumbs, GlobalHandlers, LinkedErrors, TryCatch, UserAgent } from './integrations';
 
 export const defaultIntegrations = [
@@ -144,4 +145,14 @@ export function close(timeout?: number): Promise<boolean> {
     return client.close(timeout);
   }
   return Promise.reject(false);
+}
+
+/**
+ * Wrap code within a try/catch block so the SDK is able to capture errors.
+ *
+ * @param fn A function to wrap.
+ */
+export function wrap(fn: Function): void {
+  // tslint:disable-next-line: no-unsafe-any
+  internalWrap(fn)();
 }